+++ dbm-ui/frontend/src/hooks/useTicketCloneInfo/generateCloneData/tendbcluster/rollback.ts
@@ -10,13 +10,15 @@
  * on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for
  * the specific language governing permissions and limitations under the License.
  */
-import type { MySQLRollbackDetails } from '@services/model/ticket/details/mysql';
+import type { SpiderRollbackDetails } from '@services/model/ticket/details/spider';
 import TicketModel from '@services/model/ticket/ticket';
+
+import type { BackupSources, BackupTypes } from '@views/spider-manage/rollback/pages/page1/const';
 
 import { random } from '@utils';
 
 // Spider 定点构造
-export function generateSpiderRollbackCloneData(ticketData: TicketModel<MySQLRollbackDetails>) {
+export function generateSpiderRollbackCloneData(ticketData: TicketModel<SpiderRollbackDetails>) {
   const { clusters, infos, rollback_cluster_type } = ticketData.details;
   const tableDataList = infos.map((item) => ({
     rowKey: random(),
@@ -28,8 +30,8 @@ export function generateSpiderRollbackCloneData(ticketData: TicketModel<MySQLRol
     },
     targetClusterId: item.target_cluster_id,
     rollbackHost: item.rollback_host,
-    backupSource: 'remote',
-    rollbackType: item.rollback_type,
+    backupSource: 'remote' as BackupSources,
+    rollbackType: item.rollback_type as BackupTypes,
     backupid: item.backupinfo.backup_id,
     rollbackTime: item.rollback_time,
     databases: item.databases,
