+++ dbm-ui/frontend/src/locales/zh-cn.json
@@ -3502,5 +3502,6 @@
     "单元格可同时输入多个对象，使用换行，空格或；，｜分隔，按 Enter 或失焦完成内容输入": "单元格可同时输入多个对象，使用换行，空格或；，｜分隔，按 Enter 或失焦完成内容输入",
     "包含通配符时, 每一单元格只允许输入单个对象。% 不能独立使用， * 只能单独使用": "包含通配符时, 每一单元格只允许输入单个对象。% 不能独立使用， * 只能单独使用",
     "查看域名/IP对应关系": "查看域名/IP对应关系",
+    "请先选择待回档集群": "请先选择待回档集群",
     "这行勿动！新增翻译请在上一行添加！": ""
 }
