+++ dbm-ui/frontend/src/services/source/fixpointRollback.ts
@@ -118,12 +118,5 @@ export function queryLatesBackupLog(params: {
   rollback_time: string;
   job_instance_id?: number;
 }) {
-  return http.get<
-    | {
-        backup_logs: Array<any>;
-        job_status: string;
-        message: string;
-      }
-    | BackupLogRecord
-  >(`${path}/query_latest_backup_log/`, params);
+  return http.get<BackupLogRecord>(`${path}/query_latest_backup_log/`, params);
 }
