+++ dbm-ui/frontend/src/views/mysql/rollback/pages/page1/const/backupType.ts
@@ -1,0 +1,17 @@
+import { t } from '@locales/index';
+
+export enum BackupTypes {
+  BACKUPID = 'BACKUPID',
+  TIME = 'TIME',
+}
+
+export const backupTypeList = [
+  {
+    value: BackupTypes.BACKUPID,
+    label: t('备份记录'),
+  },
+  {
+    value: BackupTypes.TIME,
+    label: t('回档到指定时间'),
+  },
+];
