+++ dbm-ui/frontend/src/views/spider-manage/rollback/pages/page1/const/backupSource.ts
@@ -1,0 +1,12 @@
+import { t } from '@locales/index';
+
+export enum BackupSources {
+  REMOTE = 'remote',
+}
+
+export const backupSourceList = [
+  {
+    value: BackupSources.REMOTE,
+    label: t('远程备份'),
+  },
+];
