+++ dbm-ui/frontend/src/views/spider-manage/rollback/pages/page1/const/index.ts
@@ -1,0 +1,2 @@
+export * from './backupSource';
+export * from '@views/mysql/rollback/pages/page1/const/backupType';
