+++ dbm-ui/frontend/src/views/tickets/common/components/demand-factory/mysql/rollback/components/RollbackExistCluster.vue
@@ -28,7 +28,7 @@
 
   import { useGlobalBizs } from '@stores';
 
-  import { type BackupSources, selectList } from '@views/mysql/rollback/pages/page1/components/common/const';
+  import { backupSourceList, type BackupSources } from '@views/mysql/rollback/pages/page1/const';
 
   import { utcDisplayTime } from '@utils';
 
@@ -58,7 +58,7 @@
       label: t('备份源'),
       field: 'backup_source',
       width: 100,
-      render: ({ cell }: { cell: BackupSources }) => <span>{ selectList.backupSource.find(item => item.value === cell)?.label || '--' }</span>,
+      render: ({ cell }: { cell: BackupSources }) => <span>{ backupSourceList.find(item => item.value === cell)?.label || '--' }</span>,
     },
     {
       label: t('回档类型'),
